kmg2/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-restx==0.5.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0
